Commit for ex 4.21
Add token authentication before deleting a blog

+++ controllers/blogs.js
@@ -25,11 +25,28 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
+// Delete only when authorized
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
-})
+  // Authentication
+  const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
 
+  const user = await User.findById(decodedToken.id)
+  const userid = user._id
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user.toString() === userid.toString() ){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else{
+    return response.status(401).json({ error: 'user is not authorized to delete this blog' })
+  }
+
+})
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
